Changed Html5HistoryRouter so that everyone shares a single instance
This means that link elements may be created that can send navigation
events to the html5 history handler without having to be given an
explicit reference in their props. This is a trade-off because those
link elements may no longer be PureComponents since they access external
state but I think that it's worth it for the practical benefits.