Changed RoutingStoreActivatorContainer to re-dispatch navigation messages to workaround async SetState behaviour
This behaviour is the default but may be disabled by passing false for
the RoutingStoreActivatorContainer's
redispatchNavigationMessagesAfterStateAsynchronouslyUpdated constructor
argument. Because SetState is (sometimes) processed asynchronously, if
navigation actions are used to change the currently-mounted container
component AND by stores to initiate data retrievals and if the data
retrieval completes before the container is mounted and subscribes to
the store's Change event then the component may miss the Change event
fired by the store and so may not realise that data is available when it
renders. Re-issueing navigation messages after the (potentially async)
SetState work has been completed means that the store can fire its
Change event again AFTER the container is known to have mounted, which
means that the container will definitely receive the event.